Auto Script Commit

+++ src/main/scala/LOFClass.scala
@@ -52,8 +52,8 @@ class LOFClass () {
 		def maxUDF=udf((neighborVal:Seq[Row]) => 
 			{
 				val returnArray = Array()
-				neighborVal.foreach(x:(Long,Double)=>
-					Math.max(10,x(1).toDouble)
+				neighborVal.foreach(x:(Long,Double) =>
+					Math.max(10,x(1))
 					)
 				println(neighborVal)
 				neighborVal
